Refactor database schema into individual files
