+++ src/database/schema/item_line.rs
@@ -1,0 +1,8 @@
+#[derive(Clone)]
+pub struct ItemLineRow {
+    pub id: String,
+    pub item_id: String,
+    pub store_id: String,
+    pub batch: String,
+    pub quantity: f64,
+}
