+++ src/database/schema/name.rs
@@ -1,0 +1,5 @@
+#[derive(Clone)]
+pub struct NameRow {
+    pub id: String,
+    pub name: String,
+}
