+++ src/database/schema/requisition_line.rs
@@ -1,0 +1,8 @@
+#[derive(Clone)]
+pub struct RequisitionLineRow {
+    pub id: String,
+    pub requisition_id: String,
+    pub item_id: String,
+    pub actual_quantity: f64,
+    pub suggested_quantity: f64,
+}
