+++ src/database/schema/store.rs
@@ -1,0 +1,5 @@
+#[derive(Clone)]
+pub struct StoreRow {
+    pub id: String,
+    pub name_id: String,
+}
