+++ src/database/schema/user_account.rs
@@ -1,0 +1,7 @@
+#[derive(Clone)]
+pub struct UserAccountRow {
+    pub id: String,
+    pub username: String,
+    pub password: String,
+    pub email: Option<String>,
+}
